refactor: Simplify ImageCard component by removing unnecessary props and handlers
